WIP: Running locally
- Following guideline at: https://github.com/Azure-Samples/guidance-identity-management-for-multitenant-apps/blob/master/docs/running-the-app.md
- Removed Startup.cs from Data project as is not an entry point (was required to exists at pre 1.0 Core era)
- REVIEW: Specified maxLength at Migrations/20151208175616_Initial.cs for those string columns that are part of an Index
- Fixed some casing for namespaces @ webapi
- Fixed ASPNETCORE_ environment variables
