+++ src/Tailspin.Surveys.Web/Properties/launchSettings.json
@@ -13,13 +13,15 @@
       "launchBrowser": true,
       "launchUrl": "https://localhost:44300/",
       "environmentVariables": {
-        "ASPNET_ENV": "Development"
+        "ASPNETCORE_ENVIRONMENT": "Development"
       }
     },
-    "web": {
-      "commandName": "web",
+    "Tailspin.Surveys.Web": {
+      "commandName": "Project",
+      "launchBrowser": true,
+      "launchUrl": "http://localhost:5000",
       "environmentVariables": {
-        "Hosting:Environment": "Development"
+        "ASPNETCORE_ENVIRONMENT": "Development"
       }
     }
   }
