+++ src/Tailspin.Surveys.Web/Tailspin.Surveys.Web.xproj
@@ -9,7 +9,7 @@
     <ProjectGuid>5024c26e-e7bb-4bed-b86b-548b8b29a630</ProjectGuid>
     <RootNamespace>Tailspin.Surveys.Web</RootNamespace>
     <BaseIntermediateOutputPath Condition="'$(BaseIntermediateOutputPath)'=='' ">..\..\artifacts</BaseIntermediateOutputPath>
-    <OutputPath Condition="'$(OutputPath)'=='' ">..\..\artifacts</OutputPath>
+    <OutputPath Condition="'$(OutputPath)'=='' ">.\bin\</OutputPath>
   </PropertyGroup>
   <PropertyGroup>
     <SchemaVersion>2.0</SchemaVersion>
