+++ src/Tailspin.Surveys.WebAPI/Properties/launchSettings.json
@@ -13,7 +13,7 @@
       "launchBrowser": true,
       "launchUrl": "https://localhost:44301/surveys/published",
       "environmentVariables": {
-        "ASPNET_ENV": "Development"
+        "ASPNETCORE_ENVIRONMENT": "Development"
       }
     },
     "web": {
