+++ src/Tailspin.Surveys.WebAPI/project.json
@@ -21,7 +21,8 @@
     "Microsoft.IdentityModel.Clients.ActiveDirectory": "3.5.208051316-alpha",
     "Tailspin.Surveys.Common": "1.0.0-*",
     "Tailspin.Surveys.Security": "1.0.0-*",
-    "Microsoft.AspNetCore.Server.IISIntegration": "1.0.0"
+    "Microsoft.AspNetCore.Server.IISIntegration": "1.0.0",
+    "Microsoft.Extensions.Configuration.CommandLine": "1.0.0"
   },
 
   "tools": {
